Fix numeric fields (#45)
The unit_field should be a text_field combined with the autonumeric
controller. This ensures appropriate formatting of the unit.

Autonumeric doesn't work with fields of the type 'number' since they
include characters like the comma.

This also fixes the sliders on the Financial Freedom Calculator since
they expect fields that have the autonumeric controller to work.